+++ ui/src/lib/person-helpers.js
@@ -113,12 +113,38 @@ export const PERSON_FRAGMENT = gql`
     headerImage { uri }
   }
 `
+export const SETTINGS_FRAGMENT = gql`
+  ${AUTHOR_FRAGMENT}
+  fragment SettingsFragment on Settings {
+    id
+    keyBackedUp
+    tombstone {
+      date
+      reason
+    }
+    recps
+    authors {
+      ...AuthorFragment
+      profile {
+        type
+        ...ProfileFragment
+      }
+    }
+    tiaki {
+      type
+      ...ProfileFragment
+    }
+    canEdit
+    originalAuthor
+  }
+`
 
 export const whoami = ({
   query: gql`
     ${PublicProfileFieldsFragment}
     ${PERSON_FRAGMENT}
     ${AUTHOR_FRAGMENT}
+    ${SETTINGS_FRAGMENT}
     query {
       whoami {
         public {
@@ -141,6 +167,9 @@ export const whoami = ({
                 ...ProfileFragment
               }
             }
+          }
+          settings {
+            ...SettingsFragment
           }
         }
       }